+++ screenshot_tool.py
@@ -1,44 +1,69 @@
-import screenshot_tool  # 修改导入路径
-import ocr_recognition
-import tkinter as tk
-from tkinter import messagebox
-import keyboard  # 导入 keyboard 库
+import pyautogui
+import keyboard
+import time
+from datetime import datetime
+from tkinter import Tk, Canvas
+import os
 
-def show_ocr_result(text):
-    def copy_to_clipboard():
-        root.clipboard_clear()
-        root.clipboard_append(text)
-        root.update()  # 现在它在剪贴板上
+def select_region():
+    root = Tk()
+    root.attributes("-fullscreen", True)
+    root.attributes("-alpha", 0.3)
+    canvas = Canvas(root, cursor="cross")
+    canvas.pack(fill="both", expand=True)
 
-    root = tk.Tk()
-    root.title("OCR 结果")
-    
-    text_box = tk.Text(root, wrap='word', height=15, width=50)
-    text_box.pack(padx=10, pady=10)
-    text_box.insert('1.0', text)
-    text_box.config(state='disabled')  # 设置为只读
+    region = []
+    rect = None
 
-    copy_button = tk.Button(root, text="复制", command=copy_to_clipboard)
-    copy_button.pack(pady=10)
+    def on_mouse_down(event):
+        nonlocal rect
+        region.append((event.x, event.y))
+        rect = canvas.create_rectangle(event.x, event.y, event.x, event.y, outline='red')
+
+    def on_mouse_move(event):
+        nonlocal rect
+        if rect:
+            canvas.coords(rect, region[0][0], region[0][1], event.x, event.y)
+
+    def on_mouse_up(event):
+        region.append((event.x, event.y))
+        root.quit()
+
+    canvas.bind("<ButtonPress-1>", on_mouse_down)
+    canvas.bind("<B1-Motion>", on_mouse_move)
+    canvas.bind("<ButtonRelease-1>", on_mouse_up)
 
     root.mainloop()
+    root.destroy()
 
-def take_screenshot_and_ocr():
-    # 截图并获取截图路径
-    screenshot_path = screenshot_tool.take_screenshot()  # 修改调用路径
-    if screenshot_path:
-        # 进行OCR识别
-        ocr_result = ocr_recognition.ocr_recognition(screenshot_path)
-        if ocr_result:
-            show_ocr_result(ocr_result)
+    if len(region) == 2:
+        x1, y1 = region[0]
+        x2, y2 = region[1]
+        # 确保坐标是正确的
+        x1, x2 = sorted([x1, x2])
+        y1, y2 = sorted([y1, y2])
+        return (x1, y1, x2 - x1, y2 - y1)
+    else:
+        return None
+
+def take_screenshot():
+    print("请拖动鼠标选择截图区域...")
+    region = select_region()
+    if region:
+        screenshot = pyautogui.screenshot(region=region)
+        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
+        os.makedirs('pix', exist_ok=True)
+        screenshot_path = f'pix/screenshot_{timestamp}.png'
+        screenshot.save(screenshot_path)
+        print(f'Screenshot saved as {screenshot_path}')
+        return screenshot_path
+    else:
+        print("未选择区域，截图取消。")
+        return None
 
 def main():
-    # 提示用户按 Alt+R 键进行截图并进行 OCR 识别
-    print("Press Alt+R to take a screenshot and perform OCR.")
-    # 绑定快捷键 Alt+R 到 take_screenshot_and_ocr 函数
-    keyboard.add_hotkey('alt+r', take_screenshot_and_ocr)
-    # 等待用户按 Esc 键退出程序
-    print("Press Esc to exit.")
+    print("Press Alt+R to take a screenshot.")
+    keyboard.add_hotkey('alt+r', take_screenshot)
     keyboard.wait('esc')
 
 if __name__ == "__main__":
